pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcription-based-summarization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "google-genai>=1.46.0",
    "litellm>=1.79.1",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "qwen-vl-utils>=0.0.14",
    "streamlit>=1.50.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
    "yt-dlp>=2025.10.22",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
